- Added default value for `file_name` parameter - All errors now are returned in JSON format - Renamed the project to innosetup_exe_builder - Added contributing and licence files - Improved Readme
