+++ README.MD
@@ -1,15 +1,21 @@
-# memori-exe-builder
+# Inno Setup-exe-builder
 
-The Memor-i-Studio project requires the ability to build a Windows executable installer. Previously, this was accomplished using wine and innoSetup on the same server. However, maintaining this setup has become increasingly difficult due to compatibility issues with newer 64-bit distributions.
+This API provides the ability to build a Windows executable installer.
 
-We created a thin flask application that serves as the builder for this app.
+Typically, this is accomplished using [Wine](https://www.winehq.org/) and [InnoSetup](https://jrsoftware.org/isinfo.php) on the same server. However, maintaining this setup has become increasingly difficult due to compatibility issues with newer 64-bit distributions.
+
+We created a thin flask application that serves as the wrapper for a Docker-based solution, and acts as a builder for .exe installers.
 
 ## Introduction
 
-This guide explains how to create a Windows executable using Inno Setup and Wine, utilizing a temporary Docker container as the builder. This functionality is utilized by our Flask application to generate the Windows executable installer at the request of our main Memor-i-Studio application.
+This guide explains how to create a Windows executable using Inno Setup and Wine, utilizing a temporary Docker container as the builder. This functionality is utilized by our Flask application to generate the Windows executable installer at the request of the client application.
 
+## Api routes exposed
 
+This api exposes a `/compile` route, which takes 2 parameters:
 
+* `path`: the absolute path of the directory that Innosetup config file resides
+* `file_name` (OPTIONAL): the file name of the Innosetup config file (.iss file)
 
 ## How to Setup/Run this app
 
@@ -26,29 +32,30 @@ To set up the environment and run the application, follow these steps:
 4. Install the required dependencies by running `pip install -r requirements.txt`.
 5. Navigate to the `project_root/src` directory.
 6. Run the application using the following command:
-    ```
+    ```bash
     gunicorn --conf gunicorn_conf.py --bind 0.0.0.0:<local-port> main:app
     ```
+7. Test a `POST` request to [http://0.0.0.0:<local-port>/compile](http://0.0.0.0:<local-port>/compile)
 
-# Read more..
-
-## Instructions to setup building exe installer manually
+## Instructions to set up building exe installer manually
 
 #### Creating Installers with Docker
 
-To create the installers for the memori game flavor, we will utilize a Docker container. This approach allows us to run commands within the container and generate the installer in the mounted directory.
+To create the installers for the required client application, we will utilize a Docker container. 
+
+This approach allows us to run commands within the container and generate the installer in the mounted directory.
 
 #### Setup Instructions
 
 1. Install Docker by following the instructions provided in the [official Docker documentation](https://docs.docker.com/engine/install/ubuntu/).
 2. Add your user to the Docker group to ensure proper permissions.
 3. Pull the required Docker image by executing the following command:
-    ```
+    ```bash
     docker pull amake/innosetup:innosetup6
     ```
 4. Test that the Docker image is working correctly by running the following commands:
-    ```
-    su - project_memori
+    ```bash
+    su - server_user
     cd /path/to/dir/with/iss-file
     docker run --rm -i -v $PWD:/work amake/innosetup:innosetup6 <iss-file-name>
     ```
@@ -57,7 +64,7 @@ To create the installers for the memori game flavor, we will utilize a Docker co
 
 #### Bash Script for Convenience
 
-```
+```text
 #!/usr/bin/env bash
 
 exec docker run --rm -i -v $PWD:/work amake/innosetup:innosetup6 "$@"
@@ -73,7 +80,7 @@ To mitigate that we could expand our script to:
 3. Create executable
 4. Change back the permissions
 
-```
+```text
 #!/usr/bin/env bash
 
 # Create the Output directory if it doesn't exist
@@ -84,3 +91,11 @@ fi
 exec docker run --rm -i -v $PWD:/work amake/innosetup:innosetup6 "$@"
 chmod 775 Output
 ```
+
+## Credits
+
+- [SciFY Dev Team](https://github.com/scify)
+
+## License
+
+The Apache Licence. Please see the [Licence File](LICENCE.md) for more information.
