+++ src/main.py
@@ -1,4 +1,4 @@
-from memori_exe_builder.app import create_app
+from innosetup_exe_builder.app import create_app
 
 app = create_app()
